requirements.txt
requests
dnspython
jinja2
